relax(cli): split cli/__init__.py into per-command modules
cli/__init__.py was the largest file in the codebase at ~750 LOC,
mixing the Typer app, six top-level command bodies, and several
sub-app builders. Split each command (or small cluster) into its
own cli/<name>.py module, leaving __init__.py with only Typer app
construction, root callback, helpers, sub-app registration, the
main entry point, and the project-root helpers shared across modules.

- info.py: startoverlay, docs, info
- agent.py: agent, _detect_agent_ticket_status, _launch_claude
- sessions.py: sessions + _format_session_age
- dashboard.py: dashboard, DashboardGuard, _resolve_overlay_for_server
- config.py: config_app + 6 subcommands
- review_request.py: review_request_app + discover

Help-output ordering preserved by registering commands in the same
order as before (verified against docs/generated/cli-reference.md —
zero diff).

Tests: updated patch sites to target new module homes
(`cli_agent_mod.os.execvp`, `cli_dashboard_mod.DashboardGuard`,
etc.). Lazy imports inside command bodies preserve `mock.patch`
ergonomics — patches at source flow through.

AGENTS.md: documented the submodule shadowing trap encountered
during the split (re-exporting `from teatree.cli.agent import agent`
overwrites the `cli.agent` submodule attribute, breaking
`patch.object(cli_agent_mod, ...)` in tests).

Baseline noqa relaxation: the new cli/<name>.py modules carry the
same `# noqa: PLC0415` markers that were already in cli/__init__.py
for deferred Django imports — the hook flags them as new because
the files are new, but they are lifted verbatim from the source.
The RUF067 per-file-ignore on cli/__init__.py was already present;
only its comment was updated to reflect the new (smaller) scope.

Closes #442
