+++ AGENTS.md
@@ -306,3 +306,4 @@ After any dashboard fix, verify the full flow before declaring done:
 - **E2E pre-flight: kill zombie servers before running.** Long-lived `uvicorn --reload` processes (from `t3 dashboard`) pin CPU at ~93% each and cause Playwright timeouts (7+ min/test). Before any e2e run: `pkill -9 -f "uvicorn teatree.asgi" 2>/dev/null; pkill -9 -f "chrome-headless" 2>/dev/null; pkill -9 -f "playwright/driver" 2>/dev/null`. Verify with `ps aux | grep uvicorn | grep -v grep`.
 - **Each timed-out e2e run leaves zombie processes.** Kill them before retrying — otherwise the next run competes for CPU with its predecessors. Same kill commands as above.
 - **E2E tests take ~7 minutes each locally on a loaded machine.** For full suite validation, prefer CI (fast, clean environment) over local runs. Run a single focused test locally to check a specific assertion, then push and let CI validate the rest.
+- **Submodule shadowing in `cli/__init__.py`.** When `cli/__init__.py` re-exports a name from a same-named submodule (`from teatree.cli.agent import agent`), the imported function overwrites the `cli.agent` submodule attribute on the parent package. Tests that do `import teatree.cli.agent as cli_agent_mod` then receive the function, not the module — `patch.object(cli_agent_mod, "os", ...)` fails with `does not have the attribute 'os'`. Use `import teatree.cli.agent as _agent` in `__init__.py` and reference attributes (`_agent.agent`) instead. The aliasing form does not bind to the parent package, so the submodule attribute survives intact.
