+++ pyproject.toml
@@ -141,7 +141,7 @@ lint.per-file-ignores."src/teatree/cli/*.py" = [
   "S607",   # partial executable paths (uv is a trusted internal tool)
 ]
 lint.per-file-ignores."src/teatree/cli/__init__.py" = [
-  "RUF067", # __init__ contains the full CLI app, not just re-exports
+  "RUF067", # Typer app construction, callbacks, helpers, and registration live here
 ]
 lint.per-file-ignores."src/teatree/core/management/commands/*.py" = [
   "PLR6301", # django-typer command methods intentionally live on Command classes
